Bundle the lex API types into a trait/struct.
The aim here is to reduce the number of type parameters that users have
to deal with: most users now only need to specify a single integer type
to get the outcome they want. Advanced users do have to deal with a bit
more boilerplate, but it's now more of a one-off, rather than requiring
lots of type parameters everywhere.

The original suggestion and implementation was from Matt Rice; I made a
second attempt; and Matt has made various additions, including the
thankless task of updating the docs. So this is definitely a co-authored
commit!

Co-authored-by: Matt Rice <ratmice@gmail.com>