Bump `regex-syntax` dependency to 0.7.
`cfgrammar`, `lrlex`, and `lrpar` pull in `regex` v1.9 which, in turn,
pulls in `regex-syntax` v0.7. `lrlex`, however, also specifies
`regex-syntax` 0.6 so that crate gets two versions built and linked in
an application.

This change was successfully run with `cargo test` so it doesn't seem
to cause a regression and now only one version of `regex-syntax` is
used.