Move generated parse function and friends to their own module.
The unsafe code test was triggering a clippy lint, in fixing that
it became apparent that unsafe code in the program section was
being denied inadvertently.

Reorganize the codegen so the generated parse tables and parse
functions are in their own `mod` separate from user actions.
That module now denies unsafe. User actions are kept in the
parent module where unsafe is still allowed.

The user provided program section is now placed at the top
of the module so it may contain it's own `#![inner_attributes]`.
This is a desired feature see #364 for one instance. While
this can be used it also applies the attribute to the generated
code in the submodule in ways that could compilation failure.

Since there is not currently a way to isolate the attributes of the
parser module from the user code section, it isn't currently a
documented or well supported feature.