Tell the user what rule is missing action code.
Before this commit a production with missing action code such as:

```
R -> <...>:
    "a" { todo!() }
  | "b"
  ;
```

would panic on an `unwrap`. Hunting down the offender could be fun!

This commit is a minimal fix for this: it tells the user that rule `R`
has a missing production. It would be better to pinpoint the exact
production, but at this point in lrpar, I think we've lost the necessary
information to do so. On the basis that something is better than
nothing, however, this commit still seems somewhat useful.