Catch incorrectly terminated productions.
Previously, if action code was followed by symbols, those would be added
to the rule, which clearly isn't what should happen. So:

```
R: A {B} C;
```

would be a rule "R" with symbols [A, C] and action code C. Whatever the
user intended to write, this probably wasn't the outcome they expected!

This commit catches this syntax error, removing a footgun.