Trim whitespace from return types.
Previously we parsed `R -> () : X` as having a return type of `() ` --
note the trailing space! This then meant that a check in `ctbuilder.rs`
that doesn't add return types if they're `()` failed, which then led to
a warning from Clippy.

Trimming the whitespace from the return type in the parser fixes this
unfortunate chain and silences Clippy.