Enforce the repair timeout when collecting repairs.
For some reason, it didn't occur to me that collecting repairs can take
a very, very long time: it's possible for this to take so long that it
exceeds any reasonable memory limit. Foolishly, I *had* realised that
`rank_rprs` needed a timeout, but I hadn't applied that to
`collect_repairs`. This commit does that, ensuring that the user's input
is less likely to cause grmtools to consume too much memory.