Add lookahead to reduce/reduce conflicts
This adds more diagnostic information to reduce/reduce conflicts based
on the observation that different lookahead symbols don't necessarily
cause the same conflicts.

For instance, the LR automaton for this grammar:

```
%start S
%%
S: A X | B X | A Y;
A: '0' ; B: '0' ; C: '0' ;
X: '1' ; Y: '2' ;
```

has the following item set for state 1:

```
1:  [B -> '0' ., {'1'}]
    [A -> '0' ., {'1', '2'}]
```

which causes a reduce/reduce conflict if the lookahead is '1', but not
if the lookahead '2'.