Add a `YaccKind::SelfDescribing` style.
This allows the yacckind value to be specified within a grammar.
When using the `YaccKind::SelfDescribing` style the grammar must
start with an initial `%grmtools` section, as documented in the variant.

For other `YaccKind` variants the `%grmtools` section is allowed (not required)
and the `YaccKind` value given to `YaccGrammar::new` overrides one given in the
`%grmtools` section.

Fixes #349