Change YaccKind parameters to a new `YaccKindResolver` type
This is a breaking change which adds a `YaccKindResolver` type.
`YaccKindResolver` can allow a yacc grammar to be read from a
`%grmtools` section of a yacc grammar source file.

CTParserBuilder which previously required a YaccKind, has
been relaxed to allow the YaccKind to be unspecified in which case
it will now attempt to find the `yacckind` in the `%grmtools` section.

nimbleparse previously would default to using `YaccKind::Original`,
now attempts to read the yacc kind from the `%grmtools` section
and otherwise throws an error. This should improve error messages
and avoid defaulting to the wrong yacc kind.

Fixes #349