Add `%grmtools` section for lrlex.
This adds a `%grmtools` section for lrlex, this enables
the ability to set `LexFlags` from within a *.l* file.

This section will be respected by `CTLexerBuilder` and tools
like *nimbleparse* and *lrlex* which read *.l* files.

For example the following lex file will change the behavior
of the "dot" regex.

```
%grmtools {!dot_matches_new_line}
%%
. "dot"
```