Migrate CACHE INFORMATION comment to a module using quote.
The quote crate doesn't deal with comments, since there isn't a
comment token.

This patch migrates the cache information to a module. The module
is nested within the `lrpar_mod!` module in order to avoid introducing
multiple modules with the same name at top-level when generating
multiple parsers.

When possible it generate constants, otherwise it generates
a function. Together these are intended to satisfy the same
purpose as the previous `CACHE INFORMATION` comment.