Replace serde requirement with bincode requirement.
Previously serde was a required dependency for the
`lrpar` crate, and the serde feature of cfgrammar was
enabled by lrpar.

Internally lrpar used the bincode serde serializers.
The bincode project now provides it's own decode
and encode traits similar to serde's serialize and
deserialize.

This changes lrpar to use the bincode traits,
making serde entirely optional.