Provide a span for `Production`
This provides a span for `Production`, even when the prod is
empty. The production will span all it's `Symbols`. Previously when
a production was empty, or `%empty` it had no Symbols and subsequently
no span. Now an empty production will at least have a zero length span
on the production.