Rework grmtools section syntax.
This reworks the syntax for the grmtools section to be unambiguous,
adding a lexer and grammar file for it in `cttests/grmtools_section.test`.
The grammar is run over the grmtools sections extracted from the .l and .y
files in the repository.

Previous unreleased versions of grmtools including this syntax used a
space separated sequence of values. Where some values were flags,
and other values acted like key value pairs.

before:
    `%grmtools{!dot_matches_new_line case_insensitive size_limit 0}`
    `%grmtools{yacckind Grmtools}`

This changes the grammar such that values must be separated by commas,
and `key: value` pairs must be separated by colons, with an optional
trailing comma.

    `%grmtools{!dot_matches_new_line, case_insensitive, size_limit: 0}`
    `%grmtools{yacckind: Grmtools}`