Add a manual implementation of the grmtools_section_parser from cttests.
This adds a `pub` parser in cfgrammar for parsing the `%grmtools` section.

It is intended to replace both the manually written parser in GrammarAST,
as well as the one in LexParser. This is also in preparation of adding
values for `RecoveryKind` and `LexerKind` to their respective sections.
Since those values live outside of `GrammarAST` and `LexParser` entirely.

The testsuite checks that the return values from both parsers are equal.
This also fixes a bug in the action code of the `grammar_section.test`
Which was return a "::" token in place of a value.