Add a manual implementation of the grmtools_section_parser from cttests.
This adds a `pub` parser in cfgrammar for parsing the `%grmtools` section.

It is intended to replace both the manually written parser in `YaccParser`,
as well as the one in `LexParser`. This is also in preparation of adding
values for `RecoveryKind` and `LexerKind` to their respective files.
Since those values live in their respective Builder.

The cttests parser action code and parser result are changed to
better match the manual parser, and their parse results checked
for equality.