Initial hacks to get merging wired up.
It contains a lot of hacks, like `unwrap()` and made up `Span::new()`.
This is because this transitions the `Header` structure from being
something that was always parsed, into something that gets constructed
by `CTParserBuilder` and then merged with what is parsed.

This is largely just an experiment with this approach to find out
if it is actually viable.