Make `YaccGrammarResult` an enum
When comparing command-line derived, or values from `CTParserBuilder`, to elements
which are parsed from an input source we may or may not have span information.

This includes an enum `Location` which contains a span, but can specify
`CommandLine`, or `Other` location. The parts of `YaccParser::parse` which
compare elements that are lacking spans, return errors that contain `Location`.

Other parts of `YaccParser` *only* deal with elements that derive from input sources.
which still return `YaccGrammarError`.  So the top-level enum changed to one which
can contain *either* `YaccGrammarError` *or* these other error kinds.