Add `Location`, `ParserError`, change `YaccParser::parse` to return ParserError.
This adds a type `Location`, for non-span specific error
locations such as `CommandLine`. In preparation for converting
command-line arguments into `Header`, and passing a header built
from the command line into the `YaccParser`.

This tries to minimize the impact to the `YaccParser` code by
keeping most of the internals of `YaccParser` still returning
`YaccGrammarError`, and converting these errors at the fringes.