Add `MarkMap`
MarkMap is a key, value map with an API similar to the typical
`std::collections` map API. The typical map API is extended in
two ways, a mechanism for marking keys with conditions. Such as
whether they a key is used and required. Additionally it has an
API for merging mark maps, and controlling merge behavior on a
per key basis.