Add a `process_header` function to `CTParserBuilder`
This function, gets a mutable reference to the `Header`,
the recovery kind, and a reference to the `YaccGrammar`.

This occurrs after the YaccGrammar is built by `build()`,
but before the `output_path` has been written.

The `process_header` function can return an error result.
Which will abort the build process.