Remove `Header` as parameters to `YaccParser` and others.
This is an interim change which removes the `Header` parameter.
It also adds `FromStr` implementations for generating grammars
without first passing in a `YaccKind`.

This patch contains a few `unwrap` calls where we currently cannot convert
into the appropriate error return type. Which will be fixed in the next patch.