Fix a panic uncovered during testing
`occupied_entry()` is documented as being for a specific use case of
mutating the flags for the *key*, using `occupied_entry()` with `insert`
runs afoul of this by assuming it can also return a value previously associated
with the key.

This is a case where the header's valueless key API doesn't coincide very
well with the existing `std::collections` entry API.  In a sense `Header`
can be occupied with a key *without* any associated value.

While for the `std::collections` entry API the existence of a key implies
the existence of a value.

This fixes the problematic usage of the API, but it would be good to revisit
the non_standard `occupied_entry()` function, and try to make another API that
is more misuse resistent such as perhaps `key_entry` which allows you to modify
the flags associated with a key, but doesn't assume the existence of a value.