Replace a misuse prone internal API with a better one.
`occupied_entry` was documented as not being useful for `insert`,
but ok for setting various marks. This is because `OccupiedEntry`
combines insertion with lookup by returning the previous value.
In doing so it assumes that the previous value exists, where in
the case of the `occupied_entry` function called from a
`VacantEntry` there was no previous value.

Instead this patch just adds the various marking functions
to `VacantEntry` avoiding the problematic assumptions.