Allow building grammars with modified start rule.
This attempts to improve the status quo regarding issue #484.
Previously in order to build a grammar using multiple start rules
you needed multiple copies of the grammar in the filesystem,
each with a different start rule.

This adds a mechanism for modifying the AST by cloning and changing
the start rules, then building a parser from each clone.
The `CTParserBuilder` methods being unstable and require
a build-dependency `feature`.