Allow using generics with `%parse-param`
This commit adds a new directive `%parse-generics`, which allows passing arbitrary generics (types and lifetimes) to the generated parser implementation.

The generics are propagated to action functions and `__GtActionsKind`, so they're available in return types as well.

This change only affects generated code: generics are not propagated to parser backend.

I've also added an example with my use case: passing an arena allocator, and returning references to AST  nodes.