Fix compile error in generated code.
`Node` is not exported by the `lrpar_mod` macro. This macro only
includes the generated Rust parser for your grammar into the module it
is called from.

Since `Node` is in the same module that lrpar generates (because it's
imported), the simplest and most correct fix is to refer to the Node
that's already in scope.

Modifying the macro to export `Node` wouldn't work because the macro
can be called from any module within a crate. For example:

```rust
// lib.rs

mod child {
    lrpar_mod!("rust.y");
}
```

would still generate an error because `crate::Node` still doesn't
exist (only `crate::child::Node` does).