Add extra_files yaml key to cttests, test %grmtools{test_files}
Because `%grmtools{test_files}` needs *both* the lexer and parser
to work. We migrate this code from using the separate `CTLexerBuilder`
and `CTParserBuilder` with `rule_ids_map` to the combined method using
`lrpar_config`.

That causes `%grmtools{test_files}` to be invoked, also add an `extra_files`
keys to the `.test` file yaml.  These emit the files to `$OUT_DIR` where the
`test_files` globs are relative to.